added configuration for color

+++ src/complex.rs
@@ -1,7 +1,7 @@
 #[derive(Clone)]
 pub struct Complex {
-    pub a : f64,
-    pub b : f64,
+    pub a : f32,
+    pub b : f32,
 }
 
 impl Complex {
@@ -12,7 +12,7 @@ impl Complex {
 
     pub fn square_it(&mut self){
         let a = self.a*self.a - self.b*self.b;
-        let b = 2f64*self.a*self.b;
+        let b = 2f32*self.a*self.b;
         self.a = a;
         self.b = b;
     }
@@ -24,13 +24,13 @@ pub struct Mandelbrot {
 }
 
 impl Mandelbrot {
-    pub fn gradient(&self, c: Complex) -> f64{
-        let mut z = Complex{a:0f64, b:0f64};
+    pub fn gradient(&self, c: Complex) -> f32{
+        let mut z = Complex{a:0f32, b:0f32};
         for i in 0..self.depth {
             z.square_it();
             z.add_to_it(&c);
-            if z.a*z.a + z.b*z.b > 5f64 {
-                return i as f64 / self.depth as f64
+            if z.a*z.a + z.b*z.b > 5f32 {
+                return i as f32 / self.depth as f32
             }
         }
         return 1.0
@@ -47,12 +47,12 @@ pub struct Julia {
 }
 
 impl Julia {
-    pub fn gradient(&self, mut z: Complex) -> f64 {
+    pub fn gradient(&self, mut z: Complex) -> f32 {
         for i in 0..self.depth {
             z.square_it();
             z.add_to_it(&self.c);
-            if z.a*z.a+z.b*z.b > 5f64{
-                return i as f64 / self.depth as f64 
+            if z.a*z.a+z.b*z.b > 5f32{
+                return i as f32 / self.depth as f32 
             }
         }
         return 1.0
